dom/phonenumberutils/tests/test_phonenumberutils_basics.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=674720
-->
<head>
  <title>Test for Bug 883923 PhoneNumberUtils FuzzyMatch</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=883923">Mozilla Bug 883923</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

"use strict";

var mozPhoneNumberService = window.navigator.mozPhoneNumberService;
if (mozPhoneNumberService) {
  is(mozPhoneNumberService.fuzzyMatch, undefined, "Fuzzy Match should not be accesible");
  is(mozPhoneNumberService.normalize, undefined, "Normalize should not be accessible");
}
</script>
</pre>
</body>
</html>
